pipeline/requirements.txt
apache-airflow[azure]==2.3.2
apache-airflow-providers-google==8.1.0
markupsafe==2.0.1
google-cloud-core==2.2.2
pytest==7.0.1
pytest-mock==3.7.0
pytest-html==3.1.1
pandas==1.4.1
pyarrow==7.0.0
mypy
protobuf==3.20.*
pytest-azurepipelines==1.0.3